+++ libs/core/src/thds/core/iterators.py
@@ -1,0 +1,9 @@
+import typing as ty
+
+T = ty.TypeVar("T")
+
+
+def null_safe_iter(it: ty.Optional[ty.Iterable[T]]) -> ty.Iterator[T]:
+    """Iterate the iterable if it is not None"""
+    if it is not None:
+        yield from it